feat: introduce listing and randomization
Introduce randomization and listing ability of the library
using the unique-random-array library. Using version
1.0.0 instead of the latest version (4.0.0) due to conflict
when trying to test it in Node according to the tutorial
instructions.

May change this in the future to follow JavaScript and
clean code best practices and to be able to use ES modules
in place of CommonJS modules.

Signed-off-by: Bethany Fannin <contact@bethanyfannin.com>
Resolves: n/a